Fix bug when saving the results
Problem: When using the timeout option, the result saving function
sometimes encountered "KeyError: 'web'".

Source: The thread modifies the result dictionnary the first time for
the lookup and the second time for the web server. With the timeout
value on the join function, the thread doesn't have time to update the
web server information.  Therefore, the key is not created.

Solution: Only modify once the result dictionnary, after the two
operations.